+++ tests/callmatrix/test_matrix.py
@@ -9,6 +9,7 @@ from httpx import AsyncClient
 from sqlmodel import select
 
 from treg import api as A
+from treg.routers import call as call_routes
 from treg import audit, ledger
 from treg.db import session_maker
 from treg.ledger import with_margin
@@ -907,7 +908,7 @@ async def test_h1_saturation_503_still_carries_a_call_id_and_a_row(
     async def _no_slot(*args, **kwargs):
         raise A.PoolTimeoutError("QueuePool limit of size 5 overflow 10 reached, connection timed out")
 
-    monkeypatch.setattr(A, "_resolve_call", _no_slot)
+    monkeypatch.setattr(call_routes, "_resolve_call", _no_slot)
     before = await snapshot(matrix_clients, fake_provider)
 
     response = await matrix_clients.get("/call/echo/ping")
