+++ tests/test_call_pool_discipline.py
@@ -26,6 +26,7 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import create_async_engine
 
 from treg import api as A
+from treg.routers import call as call_routes
 from treg import audit, ledger
 from treg.config import get_settings
 from treg.db import _engine, session_maker
@@ -62,7 +63,7 @@ async def test_a_metered_call_holds_no_db_connection_while_upstream_is_called(
 ):
     await audit.drain()  # an earlier test's fire-and-forget audit row must not be counted
     seen: list[int] = []
-    monkeypatch.setattr(A, "relay", _relay_that_checks_the_pool(seen))
+    monkeypatch.setattr(call_routes, "relay", _relay_that_checks_the_pool(seen))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 200, r.text
     assert seen == [0], f"pooled connections held during the upstream round trip: {seen}"
@@ -76,7 +77,7 @@ async def test_an_own_key_call_holds_no_db_connection_while_upstream_is_called(
     await clients.post("/secrets", json={"name": "tikhub", "value": "MKKEY"})
     await audit.drain()
     seen: list[int] = []
-    monkeypatch.setattr(A, "relay", _relay_that_checks_the_pool(seen))
+    monkeypatch.setattr(call_routes, "relay", _relay_that_checks_the_pool(seen))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 200 and r.json()["auth"] == "Bearer MKKEY"
     assert seen == [0], f"pooled connections held during the upstream round trip: {seen}"
@@ -102,7 +103,7 @@ async def test_a_burst_larger_than_the_pool_settles_every_call_at_provider_speed
         await asyncio.wait_for(everyone_in.wait(), timeout=10)
         return await original(*args, **kwargs)
 
-    monkeypatch.setattr(A, "relay", _relay_after_everyone_arrives)
+    monkeypatch.setattr(call_routes, "relay", _relay_after_everyone_arrives)
     org_id = (await clients.get("/orgs")).json()[0]["org_id"]
     before = (await clients.get(f"/orgs/{org_id}/balance")).json()["balance_micro"]
 
@@ -126,7 +127,7 @@ async def test_a_saturated_pool_answers_a_typed_503_not_an_anonymous_500(
     async def _no_slot(*args, **kwargs):
         raise PoolTimeoutError("QueuePool limit of size 5 overflow 10 reached, connection timed out")
 
-    monkeypatch.setattr(A, "_resolve_call", _no_slot)
+    monkeypatch.setattr(call_routes, "_resolve_call", _no_slot)
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 503, r.text
     assert r.json()["treg_saturated"] is True
