+++ tests/test_error_capture.py
@@ -26,6 +26,7 @@ from httpx import AsyncClient
 from sqlalchemy import select
 
 from treg import api as A
+from treg.routers import call as call_routes
 from treg.config import get_settings
 from treg.db import session_maker
 from treg.models import CallRecord
@@ -154,7 +155,7 @@ async def test_a_successful_platform_call_stores_no_evidence(clients: AsyncClien
 # ---- shape 1: the provider answered badly -------------------------------------------------------
 async def test_a_provider_failure_keeps_its_message_and_the_caller_s_query(
         clients: AsyncClient, platform_on, monkeypatch):
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         400, b'{"error":"aweme_id must be numeric","code":"E_BAD_PARAM"}'))
     r = await clients.get(f"/call/{EP}?aweme_id=not-a-number&count=5")
     assert r.status_code == 400
@@ -167,7 +168,7 @@ async def test_a_provider_failure_keeps_its_message_and_the_caller_s_query(
 async def test_a_post_body_is_captured_for_a_failed_platform_call(
         clients: AsyncClient, platform_on, monkeypatch):
     """The body is the only place a POST endpoint's parameters live — `path` never carries them."""
-    monkeypatch.setattr(A, "relay", _fake_relay(422, b'{"detail":"target is required"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(422, b'{"detail":"target is required"}'))
     r = await clients.post(f"/call/{EP_POST}", json=[{"url": "https://example.com/coffee"}])
     assert r.status_code == 422, r.text
     row = await _row(clients)
@@ -180,7 +181,7 @@ async def test_treg_s_own_502_is_explained_too(clients: AsyncClient, platform_on
     """The branch where `body` is UNBOUND. These are the failures a bare status explains least:
     upstream timeout, connection reset, failed injection, SSRF refusal."""
     import httpx
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         200, raises=httpx.ConnectTimeout("timed out after 30s")))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 502
@@ -195,7 +196,7 @@ async def test_an_own_key_failure_keeps_redacted_evidence(clients: AsyncClient, 
     """Tier 2 failures retain evidence, but an echoed org credential never survives."""
     own_key = "org.own-Q7"
     await clients.post("/secrets", json={"name": "tikhub", "value": own_key})
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         400, f'{{"error":"their own failure","received":"{own_key}"}}'.encode()))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 400
@@ -212,7 +213,7 @@ async def test_oauth_blob_masks_tokens_but_keeps_scope(clients: AsyncClient, mon
                        "scope": "webmasters.readonly", "token_type": "Bearer"})
     await _own_tool(clients, name="oauth-mask", value=blob, kind="oauth", injector="oauth",
                     secret_field="access_token")
-    monkeypatch.setattr(A, "relay", _fake_relay(403, json.dumps({
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(403, json.dumps({
         "error": "invalid grant", "scope": "webmasters.readonly",
         "access_token": access, "refresh_token": refresh, "token_type": "Bearer",
     }).encode()))
@@ -226,7 +227,7 @@ async def test_oauth_blob_masks_tokens_but_keeps_scope(clients: AsyncClient, mon
 
 async def test_plain_own_tool_failure_keeps_evidence(clients: AsyncClient, monkeypatch):
     await _own_tool(clients, name="plain-own")
-    monkeypatch.setattr(A, "relay", _fake_relay(422, b'{"error":"missing report dimensions"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(422, b'{"error":"missing report dimensions"}'))
     r = await clients.get("/call/plain-own/reports?property=123")
     assert r.status_code == 422
     row = await _row(clients)
@@ -249,7 +250,7 @@ async def test_oauth_provisioned_own_tool_failure_keeps_evidence(
     callback = await clients.get(f"/oauth/callback?code=AUTHCODE&state={state}")
     assert callback.status_code == 200, callback.text
 
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"invalid siteUrl"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"invalid siteUrl"}'))
     r = await clients.get("/call/google-search-console/sites?siteUrl=sc-domain%3Aexample.com")
     assert r.status_code == 400
     row = await _row(clients)
@@ -261,7 +262,7 @@ async def test_oauth_provisioned_own_tool_failure_keeps_evidence(
 async def test_masking_render_failure_is_redacted_not_a_500(clients: AsyncClient, monkeypatch):
     await _own_tool(clients, name="masking-fails")
     monkeypatch.setattr(A, "_secret_renderings", lambda *args: (_ for _ in ()).throw(ValueError("bad")))
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"credential echoed here"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"credential echoed here"}'))
     r = await clients.get("/call/masking-fails/fail?case=render")
     assert r.status_code == 400
     row = await _row(clients)
@@ -281,7 +282,7 @@ async def test_streaming_4xx_reaches_caller_byte_for_byte_and_keeps_evidence(
         return StreamingResponse(stream(), status_code=400,
                                  headers={"x-request-id": "req-stream-1"})
 
-    monkeypatch.setattr(A, "relay", chunked)
+    monkeypatch.setattr(call_routes, "relay", chunked)
     r = await clients.get("/call/stream-own/fail?part=all")
     assert r.status_code == 400
     assert r.content == b"".join(chunks)
@@ -292,7 +293,7 @@ async def test_streaming_4xx_reaches_caller_byte_for_byte_and_keeps_evidence(
 
 async def test_large_unmetered_body_keeps_query_only(clients: AsyncClient, monkeypatch):
     await _own_tool(clients, name="large-own")
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"too large"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"too large"}'))
     body = b"body-marker-" + b"x" * A._ERROR_CALLER_BODY_MAX
     r = await clients.post("/call/large-own/fail?request_id=query-only", content=body)
     assert r.status_code == 400
@@ -305,7 +306,7 @@ async def test_treg_side_502_keeps_own_tool_evidence(clients: AsyncClient, monke
     import httpx
 
     await _own_tool(clients, name="broken-own")
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         502, raises=httpx.ConnectError("connection reset by peer")))
     r = await clients.get("/call/broken-own/fail?request_id=502-case")
     assert r.status_code == 502
@@ -327,7 +328,7 @@ async def test_a_treg_refusal_stores_nothing(clients: AsyncClient):
 async def test_the_platform_key_never_reaches_the_columns_header_injected(
         clients: AsyncClient, platform_on, monkeypatch):
     """The realistic leak: a provider quoting the credential it received back inside its 401 body."""
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         401, f'{{"error":"invalid key","received":"Bearer {PLATFORM_KEY}"}}'.encode()))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 401
@@ -347,7 +348,7 @@ async def test_the_platform_key_never_reaches_the_columns_query_injected(
     # second copy this test passes with `_secret_renderings` disabled entirely — verified.
     echoed = (f'{{"error":"key {SPYFU_KEY} is not valid for this endpoint",'
               f'"url":"https://api.spyfu.com/x?api_key={quote(SPYFU_KEY, safe="")}&domain=a.com"}}')
-    monkeypatch.setattr(A, "relay", _fake_relay(403, echoed.encode()))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(403, echoed.encode()))
     r = await clients.get(f"/call/{EP_SPYFU}?domain=a.com")
     assert r.status_code == 403
     row = await _row(clients)
@@ -358,7 +359,7 @@ async def test_the_platform_key_never_reaches_the_columns_query_injected(
 async def test_a_callers_own_value_in_the_injected_slot_is_dropped(
         clients: AsyncClient, platform_on, monkeypatch):
     """A caller who passes their own value into the param the injector overwrites."""
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"bad request"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"bad request"}'))
     r = await clients.get(f"/call/{EP_SPYFU}?domain=a.com&api_key=CALLERS-OWN-SECRET-VALUE")
     assert r.status_code == 400
     row = await _row(clients)
@@ -377,7 +378,7 @@ async def test_a_decoded_basic_credential_never_survives(clients: AsyncClient, p
     monkeypatch.setenv("TREG_PLATFORM_KEY_DATAFORSEO",
                        base64.b64encode(f"{login}:{password}".encode()).decode())
     get_settings.cache_clear()
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         401, json.dumps({"error": "auth failed",
                          "received_username": login,
                          "received_password": password}).encode()))
@@ -395,7 +396,7 @@ async def test_a_lowercase_percent_encoded_key_never_survives(clients: AsyncClie
     from urllib.parse import quote
     monkeypatch.setenv("TREG_PLATFORM_KEY_TIKHUB", "a/b+c=dQ")
     get_settings.cache_clear()
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         400, f'{{"error":"bad","echo":"{quote("a/b+c=dQ", safe="").lower()}"}}'.encode()))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 400
@@ -412,7 +413,7 @@ async def test_an_empty_bodied_429_still_says_when_to_retry(clients: AsyncClient
                                                             monkeypatch):
     """The real tikhub population is 70 401s and 54 429s. Those bodies are often empty or generic,
     and the headers are then the entire diagnosis — 'quota gone, back in 60s' versus 'bad key'."""
-    monkeypatch.setattr(A, "relay", _fake_relay(429, b"", headers={
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(429, b"", headers={
         "retry-after": "60", "x-ratelimit-remaining": "0", "x-request-id": "req-8f2a4c19"}))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 429
@@ -424,7 +425,7 @@ async def test_an_empty_bodied_429_still_says_when_to_retry(clients: AsyncClient
 
 async def test_a_401_keeps_the_auth_challenge_and_not_the_credential(clients: AsyncClient,
                                                                      platform_on, monkeypatch):
-    monkeypatch.setattr(A, "relay", _fake_relay(401, b"", headers={
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(401, b"", headers={
         "www-authenticate": 'Bearer realm="api", error="invalid_token"'}))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 401
@@ -439,7 +440,7 @@ async def test_a_provider_correlation_id_survives_redaction(clients: AsyncClient
     so UUIDs, trace ids and request ids were deleted 100% of the time — measured on real bodies, the
     prose always survived and the correlation field never did."""
     trace = "11da3d88-e351-4c07-87ea-f5160d76a87d"          # 36 chars: the old rule ate this
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         400, f'{{"message":"bad keyword","request_id":"{trace}"}}'.encode()))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 400
@@ -458,7 +459,7 @@ async def test_a_real_secret_shape_is_still_masked(clients: AsyncClient, platfor
     """
     fake_key = "sk_live_ABCDEFGHIJKLMNOP1234"
     fake_jwt = "eyJhbGciOi.JIUzI1NiIsInR5cCI6"
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         400, f'{{"message":"bad token {fake_key} and {fake_jwt}"}}'.encode()))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 400
@@ -471,7 +472,7 @@ async def test_a_bodyless_failure_still_leaves_a_row(clients: AsyncClient, platf
     """A 4xx with no body and none of the allowlisted headers used to produce "" for both snippets,
     and `_audit` only stores evidence when one is truthy — so the row vanished from /admin/errors and
     the failure looked like capture had never run. "Nothing came back" is itself the finding."""
-    monkeypatch.setattr(A, "relay", _fake_relay(500, b""))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(500, b""))
     # The required param has to be present or treg refuses before relay — the empty half under test
     # is the RESPONSE, which is what a bodyless provider 5xx actually looks like.
     r = await clients.get(f"/call/{EP}?aweme_id=7")
@@ -485,7 +486,7 @@ async def test_a_bodyless_failure_still_leaves_a_row(clients: AsyncClient, platf
 async def test_expired_evidence_is_a_state_not_content(clients: AsyncClient, platform_on,
                                                        monkeypatch):
     """The sentinel must never be served as though it were the provider's answer."""
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"aged out later"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"aged out later"}'))
     await clients.get(f"/call/{EP}?aweme_id=7")
     from treg import audit
     await audit.drain()
@@ -510,7 +511,7 @@ async def test_a_cap_refusal_says_WHICH_cap(clients: AsyncClient, platform_on, m
     async def _boom(*a, **k):
         raise HTTPException(status_code=429, detail="daily call cap reached for this member (500/500)")
 
-    monkeypatch.setattr(A, "_platform_reserve", _boom)
+    monkeypatch.setattr(call_routes, "_platform_reserve", _boom)
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 429
     row = await _row(clients)
@@ -520,7 +521,7 @@ async def test_a_cap_refusal_says_WHICH_cap(clients: AsyncClient, platform_on, m
 
 async def test_admin_errors_surfaces_the_method(clients: AsyncClient, platform_on, monkeypatch):
     """A GET at a POST endpoint IS the diagnosis for 47 real apollo failures."""
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"nope"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"nope"}'))
     await clients.get(f"/call/{EP}?aweme_id=7")
     from treg import audit
     await audit.drain()
@@ -533,7 +534,7 @@ async def test_a_gzipped_error_page_does_not_become_replacement_characters(
         clients: AsyncClient, platform_on, monkeypatch):
     """`force_identity` asks a provider not to compress, but a CDN/WAF error page is generated at the
     edge and answers however it likes — and those 403s are exactly what this feature is for."""
-    monkeypatch.setattr(A, "relay", _fake_relay(
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(
         403, gzip.compress(b'{"error":"blocked by firewall, ray id 8f2a"}'),
         headers={"content-encoding": "gzip"}))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
@@ -544,7 +545,7 @@ async def test_a_gzipped_error_page_does_not_become_replacement_characters(
 
 
 async def test_a_binary_body_is_described_not_mangled(clients: AsyncClient, platform_on, monkeypatch):
-    monkeypatch.setattr(A, "relay", _fake_relay(500, b"\x89PNG\r\n\x1a\n\x00\x00\x00\x00" * 8))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(500, b"\x89PNG\r\n\x1a\n\x00\x00\x00\x00" * 8))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 500
     row = await _row(clients)
@@ -553,7 +554,7 @@ async def test_a_binary_body_is_described_not_mangled(clients: AsyncClient, plat
 
 async def test_a_huge_error_page_is_truncated(clients: AsyncClient, platform_on, monkeypatch):
     huge = b"<html><body>" + b"the server encountered an error. " * 2000 + b"</body></html>"
-    monkeypatch.setattr(A, "relay", _fake_relay(500, huge))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(500, huge))
     r = await clients.get(f"/call/{EP}?aweme_id=7")
     assert r.status_code == 500
     row = await _row(clients)
@@ -594,7 +595,7 @@ async def test_the_columns_are_not_exposed_to_the_team_yet(clients: AsyncClient,
                                                            monkeypatch):
     """v1 is admin-only. `/calls` builds an explicit field list, so a new column cannot appear there
     by accident — this pins that, because the follow-on that exposes it must be deliberate."""
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"nope"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"nope"}'))
     await clients.get(f"/call/{EP}?aweme_id=7")
     from treg import audit
     await audit.drain()
@@ -605,18 +606,18 @@ async def test_the_columns_are_not_exposed_to_the_team_yet(clients: AsyncClient,
 # ---- the admin view, and ageing ------------------------------------------------------------------
 async def test_admin_errors_lists_all_failed_tiers_and_filters_them(clients: AsyncClient, platform_on,
                                                                    monkeypatch):
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"aweme_id must be numeric"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"aweme_id must be numeric"}'))
     await clients.get(f"/call/{EP}?aweme_id=bad")
 
     await clients.post("/secrets", json={"name": "tikhub", "value": "org.own-Q7"})
-    monkeypatch.setattr(A, "relay", _fake_relay(401, b'{"error":"own key rejected"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(401, b'{"error":"own key rejected"}'))
     await clients.get(f"/call/{EP}?aweme_id=own")
 
     await _own_tool(clients, name="admin-own")
-    monkeypatch.setattr(A, "relay", _fake_relay(422, b'{"error":"own tool rejected"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(422, b'{"error":"own tool rejected"}'))
     await clients.get("/call/admin-own/fail?case=own-tool")
 
-    monkeypatch.setattr(A, "relay", _fake_relay(200, b'{"ok":true}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(200, b'{"ok":true}'))
     await clients.get("/call/admin-own/success")          # a success must not appear
     from treg import audit
     await audit.drain()
@@ -637,7 +638,7 @@ async def test_evidence_ages_out_but_the_audit_row_survives(clients: AsyncClient
                                                             monkeypatch):
     """Retention blanks the two columns and touches nothing else — the call itself is the audit
     trail and has to outlive its evidence."""
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"stale failure"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"stale failure"}'))
     await clients.get(f"/call/{EP}?aweme_id=bad")
     from treg import audit
     await audit.drain()
