+++ tests/test_tag_billing.py
@@ -19,6 +19,7 @@ from httpx import AsyncClient
 from sqlmodel import select
 
 from treg import api as A, audit, crypto, ledger
+from treg.routers import call as call_routes
 from treg.config import get_settings
 from treg.db import session_maker
 from treg.models import Membership, Org, TagSpend, User
@@ -205,7 +206,7 @@ async def test_untagged_metered_calls_write_no_tag_rows(clients: AsyncClient, pl
 async def test_a_released_call_bills_no_tag(clients: AsyncClient, platform_on, monkeypatch):
     """A provider 5xx releases the hold in full, so the builder's user must not be billed for it —
     and the tag rows go with it rather than lingering as phantom spend."""
-    monkeypatch.setattr(A, "relay", _fake_relay(503, b"nope"))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(503, b"nope"))
     org_id = await _org_id(clients)
     await clients.get(f"/call/{EP}?aweme_id=7", headers={"X-Treg-Meta": "customer=cust_A"})
     async with session_maker() as db:
@@ -754,7 +755,7 @@ async def test_a_provider_credential_refusal_is_never_billed(clients: AsyncClien
     before = (await clients.get(f"/orgs/{org_id}/balance")).json()["balance_micro"]
 
     for status in (401, 402, 403, 407, 408, 429):
-        monkeypatch.setattr(A, "relay", _fake_relay(status, b'{"message":"out of credits"}'))
+        monkeypatch.setattr(call_routes, "relay", _fake_relay(status, b'{"message":"out of credits"}'))
         r = await clients.get(f"/call/{EP}?aweme_id=7", headers={"X-Treg-Meta": "customer=cust_A"})
         assert r.status_code == status, r.text
         assert r.headers.get("X-Treg-Cost-Micro") == "0", f"{status} was billed"
@@ -769,7 +770,7 @@ async def test_a_caller_input_4xx_is_still_billed_on_a_per_call_endpoint(clients
     """The other half of the rule stays: the provider DOES charge for accepting a malformed request,
     so a caller's own bad input is on the caller."""
     org_id = await _org_id(clients)
-    monkeypatch.setattr(A, "relay", _fake_relay(400, b'{"error":"bad param"}'))
+    monkeypatch.setattr(call_routes, "relay", _fake_relay(400, b'{"error":"bad param"}'))
     r = await clients.get(f"/call/{EP}?aweme_id=7", headers={"X-Treg-Meta": "customer=cust_A"})
     assert r.status_code == 400
     # tikhub comments is per_success, so 400 releases; assert the RULE directly for per_call.
